requirements.txt
click==8.1.3
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
Werkzeug==2.2.0
